refactor(musd): move useMerklClaimStatus hook and corresponding tests
