+++ ui/components/app/toast-master/toast-master.js
@@ -62,7 +62,7 @@ import {
   ClaimSubmitToastType,
   StorageWriteErrorType,
 } from '../../../../shared/constants/app-state';
-import { useMerklClaimStatus } from '../musd/hooks/useMerklClaimStatus';
+import { useMerklClaimStatus } from '../../../hooks/musd/useMerklClaimStatus';
 import { getDappActiveNetwork } from '../../../selectors/dapp';
 import {
   getAccountGroupWithInternalAccounts,
