+++ ui/hooks/musd/useMerklClaimStatus.ts
@@ -4,7 +4,7 @@ import {
   TransactionStatus,
   type TransactionMeta,
 } from '@metamask/transaction-controller';
-import { getTransactions } from '../../../../selectors/transactions';
+import { getTransactions } from '../../selectors/transactions';
 
 export const MERKL_DISTRIBUTOR_ADDRESS =
   '0x3Ef3D8bA38EBe18DB133cEc108f4D14CE00Dd9Ae' as const;
